Customize main app button example code
This doc page helps the dev overcome using the arcane Ribbon Editor collection features. But another place where dev churn occurs is when you really try to customize the element list. As it is now, it takes the the un-initiated dev lots of time and mistakes and searches to find a good approach for doing this. Therefore, I propose that it would cut down learning time considerably by adding an Example. The example I have added adds an element button to the list where you manually have to set its data items.